created new file for AdEventType enum
