+++ lib/src/enums/ad_event_type.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/foundation.dart';
+
+enum AdEventType {
+  adNetworkInitialized,
+  adLoaded,
+  adDismissed,
+  adShowed,
+  adFailedToLoad,
+  adFailedToShow,
+  earnedReward,
+}
+
+extension AdEventTypeExtension on AdEventType {
+  String get value => describeEnum(this);
+}
