+++ lib/src/enums/ad_unit_type.dart
@@ -6,6 +6,6 @@ enum AdUnitType {
   rewarded,
 }
 
-extension AdNetworkExtension on AdUnitType {
+extension AdUnitTypeExtension on AdUnitType {
   String get value => describeEnum(this);
 }
