+++ lib/src/utils/ad_event.dart
@@ -1,15 +1,6 @@
+import 'package:easy_ads_flutter/src/enums/ad_event_type.dart';
 import 'package:easy_ads_flutter/src/enums/ad_network.dart';
 import 'package:easy_ads_flutter/src/enums/ad_unit_type.dart';
-
-enum AdEventType {
-  adNetworkInitialized,
-  adLoaded,
-  adDismissed,
-  adShowed,
-  adFailedToLoad,
-  adFailedToShow,
-  earnedReward,
-}
 
 class AdEvent {
   final AdEventType type;
